Fix: fix some clang-tidy warnings

+++ src/lua_mympd_state.c
@@ -18,7 +18,7 @@ void set_lua_mympd_state_p(struct list *lua_mympd_state, const char *k, const ch
     list_push(lua_mympd_state, k, LUA_TYPE_STRING, NULL, value);
 }
 
-void set_lua_mympd_state_i(struct list *lua_mympd_state, const char *k, int v) {
+void set_lua_mympd_state_i(struct list *lua_mympd_state, const char *k, long v) {
     struct t_lua_mympd_state_value *value = (struct t_lua_mympd_state_value *)malloc(sizeof(struct t_lua_mympd_state_value));
     value->i = v;
     list_push(lua_mympd_state, k, LUA_TYPE_INTEGER, NULL, value);
